Convert core JS files into modules.
