+++ web/doc.html
@@ -57,8 +57,7 @@
       charset="UTF-8"
     ></script>
     <script type="text/javascript" src="js/codeworld-mode.js"></script>
-    <script type="text/javascript" src="js/codeworld_shared.js"></script>
-    <script type="text/javascript" src="js/doc.js"></script>
+    <script type="module" src="js/doc.js"></script>
   </head>
 
   <body>
