+++ web/env.html
@@ -45,40 +45,22 @@
       <div class="ui-layout-container">
         <div id="nav" class="dropbox ui-layout-west">
           <div class="nav_items">
-            <a id="newFolderButton" class="cw-button red" onclick="newFolder()">
+            <a id="newFolderButton" class="cw-button red">
               <i class="mdi mdi-18px mdi-folder"></i>&nbsp; New Folder
             </a>
-            <a id="newButton" class="cw-button red" onclick="newProject()">
+            <a id="newButton" class="cw-button red">
               <i class="mdi mdi-18px mdi-plus"></i>&nbsp; New
             </a>
-            <a
-              id="saveButton"
-              class="cw-button blue"
-              style="display: none;"
-              onclick="saveProject()"
-            >
+            <a id="saveButton" class="cw-button blue" style="display: none;">
               <i class="mdi mdi-18px mdi-cloud-upload"></i>&nbsp; Save
             </a>
-            <a
-              id="saveAsButton"
-              class="cw-button blue"
-              onclick="saveProjectAs()"
-            >
+            <a id="saveAsButton" class="cw-button blue">
               <i class="mdi mdi-18px mdi-dots-horizontal"></i>&nbsp; Save As
             </a>
-            <a
-              id="deleteButton"
-              class="cw-button red"
-              style="display: none;"
-              onclick="deleteProject()"
-            >
+            <a id="deleteButton" class="cw-button red" style="display: none;">
               <i class="mdi mdi-18px mdi-delete"></i>&nbsp; Delete
             </a>
-            <a
-              id="downloadButton"
-              class="cw-button blue"
-              onclick="downloadProject()"
-            >
+            <a id="downloadButton" class="cw-button blue">
               <i class="mdi mdi-18px mdi-download"></i>&nbsp; Download
             </a>
           </div>
@@ -111,17 +93,13 @@
             onclick="window.mainLayout.toggle('west')"
             ><i class="mdi mdi-18px mdi-menu"></i
           ></a>
-          <a id="signin" class="cw-button blue" onclick="signin()"
+          <a id="signin" class="cw-button blue"
             ><i class="mdi mdi-18px mdi-login"></i>&nbsp; Sign In</a
           >
-          <a
-            id="signout"
-            class="cw-button blue"
-            style="display: none;"
-            onclick="signout()"
+          <a id="signout" class="cw-button blue" style="display: none;"
             ><i class="mdi mdi-18px mdi-logout"></i>&nbsp; Sign Out</a
           >
-          <a id="docButton" class="cw-button blue" onclick="help()"
+          <a id="docButton" class="cw-button blue"
             ><i class="mdi mdi-18px mdi-book-open-variant"></i>&nbsp; Guide</a
           >
           <a
@@ -132,10 +110,10 @@
             ><i class="mdi mdi-18px mdi-bug"></i>&nbsp; Report a Bug</a
           >
           <a
+            id="toggleThemeButton"
             class="cw-button blue theme-marker-day"
             title="Toggle theme"
             target="_blank"
-            onclick="toggleTheme()"
             ><i class="mdi mdi-18px mdi-weather-sunny"></i
           ></a>
         </div>
@@ -145,7 +123,6 @@
             id="shareFolderButton"
             class="cw-button yellow"
             style="display: none;"
-            onclick="shareFolder()"
             ><i class="mdi mdi-18px mdi-folder-outline"></i>&nbsp; Share
             Folder</a
           >
@@ -158,38 +135,22 @@
               <!--Recording Icon--></i
             ></span
           >
-          <a
-            id="startRecButton"
-            style="display: none;"
-            class="cw-button red"
-            onclick="captureStart()"
+          <a id="startRecButton" style="display: none;" class="cw-button red"
             ><i class="mdi mdi-18px mdi-record"></i>&nbsp; Start Recording</a
           >
-          <a
-            id="stopRecButton"
-            style="display: none;"
-            class="cw-button yellow"
-            onclick="stopRecording()"
+          <a id="stopRecButton" style="display: none;" class="cw-button yellow"
             ><i class="mdi mdi-18px mdi-stop"></i>&nbsp; Stop Recording</a
           >
-          <a
-            id="shareButton"
-            class="cw-button yellow"
-            style="display: none;"
-            onclick="share()"
+          <a id="shareButton" class="cw-button yellow" style="display: none;"
             ><i class="mdi mdi-18px mdi-share"></i>&nbsp; Share</a
           >
-          <a
-            id="inspectButton"
-            class="cw-button cyan"
-            style="display: none;"
-            onclick="inspect()"
+          <a id="inspectButton" class="cw-button cyan" style="display: none;"
             ><i class="mdi mdi-18px mdi-magnify"></i>&nbsp; Inspect</a
           >
-          <a class="cw-button red" onclick="stopRun()"
+          <a id="stopCompilationButton" class="cw-button red"
             ><i class="mdi mdi-18px mdi-stop"></i>&nbsp; Stop</a
           >
-          <a class="cw-button green" onclick="compile()"
+          <a id="startCompilationButton" class="cw-button green"
             ><i class="mdi mdi-18px mdi-play"></i>&nbsp; Run</a
           >
         </div>
@@ -270,18 +231,11 @@
       type="text/javascript"
       src="js/details-element-polyfill.js"
     ></script>
-    <script type="text/javascript" src="js/utils/index.js"></script>
-    <script type="text/javascript" src="js/utils/directoryTree.js"></script>
     <script type="text/javascript" src="js/codeworld-mode.js"></script>
-    <script type="text/javascript" src="js/codeworld_shared.js"></script>
-    <script type="text/javascript" src="js/codeworld.js"></script>
-    <script type="text/javascript" src="js/treedialog.js"></script>
+    <script type="module" src="js/codeworld.js"></script>
     <script
       type="text/javascript"
       src="https://cdn.jsdelivr.net/npm/promise-polyfill@7.1.0/dist/promise.min.js"
     ></script>
-    <script type="text/javascript">
-      init();
-    </script>
   </body>
 </html>
