+++ web/js/codeworld_shared.js
@@ -13,64 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-'use strict';
 
-/*
- * Utility function for sending an HTTP request to fetch a resource.
- *
- * Args:
- *   - method: The HTTP method to use, such as 'GET'
- *   - url: The URL to fetch, whether absolute or relative.
- *   - body: The request body to send.  Use null for no body.
- *   - callback: A callback function to send when complete.  (optional)
- *
- * If provided, the callback will be given the XmlHttpRequest object, so
- * it can inspect the response code and headers as well as the contents.
- */
-function sendHttp(method, url, body, callback) {
-  const sendHttpFunc = signedIn() ? window.auth2.sendHttpAuth : sendHttpRaw;
-  return sendHttpFunc(method, url, body, callback);
-}
+import * as Alert from './utils/alert.js';
+import * as Auth from './utils/auth.js';
+import * as DirTree from './utils/directoryTree.js';
+import * as Html from './utils/html.js';
+import { sendHttp } from './utils/network.js';
 
-function sendHttpRaw(method, url, body, callback) {
-  const request = new XMLHttpRequest();
-
-  if (callback) {
-    request.onreadystatechange = () => {
-      if (request.readyState === 4) callback(request);
-    };
-  }
-
-  request.open(method, url, true);
-  request.send(body);
-
-  return request;
-}
-
-const Html = (() => {
-  const mine = {};
-
-  mine.encode = (str) => $('<div/>').text(str).html();
-
-  return mine;
-})();
-
-const Alert = (() => {
-  const mine = {};
-
-  mine.init = () =>
-    Promise.resolve(
-      $.getScript(
-        'https://cdnjs.cloudflare.com/ajax/libs/limonte-sweetalert2/7.19.2/sweetalert2.all.min.js'
-      )
-    ).catch((e) => console.log('Alert.init failed'));
-
-  // Build SweetAlert title HTML
-  mine.title = (text, iconClass) =>
-    `<i class="mdi mdi-72px ${iconClass}"></i>&nbsp; ${Html.encode(text)}`;
-
-  return mine;
-})();
+import * as TreeDialog from './treedialog.js';
 
 const hintBlacklist = [
   // Symbols that only exist to implement RebindableSyntax or map to
@@ -773,134 +723,6 @@ function registerStandardHints(successFunc) {
   });
 }
 
-function signin() {
-  if (window.auth2) {
-    window.auth2.signIn({
-      prompt: 'login',
-    });
-  }
-}
-
-function signout() {
-  warnIfUnsaved(() => {
-    clearWorkspace();
-    if (window.auth2) window.auth2.signOut();
-  });
-}
-
-function signedIn() {
-  return Boolean(window.auth2 && window.auth2.isSignedIn.get());
-}
-
-const Auth = (() => {
-  const mine = {};
-
-  function initLocalAuth() {
-    Promise.resolve($.getScript('js/codeworld_local_auth.js'))
-      .then(() => onAuthInitialized(LocalAuth.init()))
-      .catch((e) => console.log('initLocalAuth failed', e));
-  }
-
-  function initGoogleAuth() {
-    Promise.resolve($.getScript('https://apis.google.com/js/platform.js'))
-      .then(() =>
-        gapi.load('auth2', () =>
-          withClientId((clientId) => {
-            function sendHttpAuth(method, url, body, callback) {
-              if (body !== null && signedIn()) {
-                const idToken = window.auth2.currentUser.get().getAuthResponse()
-                  .id_token;
-                body.append('id_token', idToken);
-              }
-
-              const request = new XMLHttpRequest();
-
-              if (callback) {
-                request.onreadystatechange = () => {
-                  if (request.readyState === 4) {
-                    callback(request);
-                  }
-                };
-              }
-
-              request.open(method, url, true);
-              request.send(body);
-
-              return request;
-            }
-
-            const auth2 = Object.assign(
-              {
-                sendHttpAuth: sendHttpAuth,
-              },
-              gapi.auth2.init({
-                client_id: clientId,
-                scope: 'openid',
-                fetch_basic_profile: false,
-              })
-            );
-
-            onAuthInitialized(auth2);
-          })
-        )
-      )
-      .catch((e) => console.log('initGoogleAuth failed'));
-  }
-
-  function onAuthInitialized(auth) {
-    window.auth2 = auth;
-    window.auth2.currentUser.listen(signinCallback);
-
-    discoverProjects('');
-  }
-
-  function onAuthDisabled() {
-    window.auth2 = null;
-    document.getElementById('signin').style.display = 'none';
-    discoverProjects('');
-  }
-
-  mine.init = () =>
-    sendHttp('GET', 'authMethod', null, (resp) => {
-      if (resp.status === 200) {
-        const obj = JSON.parse(resp.responseText);
-        switch (obj.authMethod) {
-        case 'Local':
-          initLocalAuth();
-          break;
-        case 'Google':
-          initGoogleAuth();
-          break;
-        default:
-          onAuthDisabled();
-          break;
-        }
-      } else {
-        onAuthDisabled();
-      }
-    });
-
-  return mine;
-})();
-
-function withClientId(f) {
-  if (window.clientId) return f(window.clientId);
-
-  sendHttp('GET', 'clientId.txt', null, (request) => {
-    if (request.status !== 200 || request.responseText === '') {
-      sweetAlert(
-        'Oops!',
-        'Missing API client key.  You will not be able to sign in.',
-        'warning'
-      );
-      return null;
-    }
-
-    window.clientId = request.responseText.trim();
-    return f(window.clientId);
-  });
-}
-
 function loadTreeNodesAtPath(path, node, callback) {
   const data = new FormData();
   data.append('mode', window.projectEnv);
@@ -914,7 +736,7 @@ function loadTreeNodesAtPath(path, node, callback) {
 
       treeNodes.forEach((node) => {
         if (!node.id) {
-          node.id = utils.directoryTree.createNodeId(node.type, node.name);
+          node.id = DirTree.createNodeId(node.type, node.name);
         }
       });
 
@@ -939,11 +761,11 @@ function loadTreeNodesAtPath(path, node, callback) {
 }
 
 function loadSubTree(node, callback) {
-  if (signedIn()) {
+  if (Auth.signedIn()) {
     // Root node already loaded
     if (node === $('#directoryTree').tree('getTree') && callback) {
       callback();
-    } else if (utils.directoryTree.isDirectory(node)) {
+    } else if (DirTree.isDirectory(node)) {
       loadTreeNodesAtPath(getNearestDirectory(node), node, callback);
     }
   } else {
@@ -952,7 +774,7 @@ function loadSubTree(node, callback) {
 }
 
 function discoverProjects(path) {
-  if (signedIn()) {
+  if (Auth.signedIn()) {
     loadTreeNodesAtPath(path);
   } else {
     updateUI();
@@ -967,7 +789,7 @@ function moveDirTreeNode(
   buildMode,
   successFunc
 ) {
-  if (!signedIn()) {
+  if (!Auth.signedIn()) {
     sweetAlert('Oops!', 'You must sign in before moving.', 'error');
     return;
   }
@@ -1023,12 +845,10 @@ function warnIfUnsaved(action) {
   if (isEditorClean()) {
     action();
   } else {
-    const msg =
-      'There are unsaved changes to your project. ' +
-      'Continue and throw away your changes?';
     sweetAlert({
       title: Alert.title('Warning'),
-      text: msg,
+      text:
+        'There are unsaved changes to your project. Continue and throw away your changes?',
       type: 'warning',
       showCancelButton: true,
       confirmButtonColor: '#DD6B55',
@@ -1039,154 +859,11 @@ function warnIfUnsaved(action) {
   }
 }
 
-function saveProjectAsBase(successFunc) {
-  if (!signedIn()) {
-    sweetAlert('Oops!', 'You must sign in to save files.', 'error');
-    updateUI();
-    return;
-  }
-
-  const selectedNode = utils.directoryTree.getSelectedNode();
-  const isDirectoryNode = utils.directoryTree.isDirectory(selectedNode);
-
-  sweetAlert({
-    title: Alert.title('Save As', 'mdi-cloud-upload'),
-    html:
-      selectedNode && isDirectoryNode
-        ? `Enter a name for your project in folder <b>${$('<div>')
-          .text(getNearestDirectory())
-          .html()
-          .replace(/ /g, '&nbsp;')}:`
-        : 'Enter a name for your project:',
-    input: 'text',
-    inputValue: selectedNode && !isDirectoryNode ? selectedNode.name : '',
-    confirmButtonText: 'Save',
-    showCancelButton: true,
-    closeOnConfirm: false,
-  }).then((result) => {
-    const parent = getNearestDirectory_();
-
-    function localSuccessFunc() {
-      const matches = parent.children.filter(
-        (node) =>
-          node.name === result.value && utils.directoryTree.isProject(node)
-      );
-      let node;
-      const type = utils.directoryTree.nodeTypes.PROJECT;
-      const name = result.value;
-
-      if (matches.length === 0) {
-        node = $('#directoryTree').tree(
-          'appendNode',
-          {
-            id: utils.directoryTree.createNodeId(type, name),
-            name,
-            type,
-            data: JSON.stringify(getCurrentProject()),
-          },
-          parent
-        );
-      } else {
-        node = matches[0];
-      }
-
-      $('#directoryTree').tree('selectNode', node);
-
-      updateChildrenIndexes(parent);
-
-      successFunc(result.value);
-    }
-    if (result && result.value) {
-      saveProjectBase(
-        getNearestDirectory(),
-        result.value,
-        window.projectEnv,
-        localSuccessFunc
-      );
-    }
-  });
-}
-
-function saveProjectBase(path, projectName, mode, successFunc) {
-  if (!signedIn()) {
-    sweetAlert('Oops!', 'You must sign in to save files.', 'error');
-    updateUI();
-    return;
-  }
-
-  if (!projectName) return;
-
-  function go() {
-    sweetAlert({
-      title: Alert.title(`Saving ${projectName} ...`),
-      text: 'Saving your project.  Please wait.',
-      showConfirmButton: false,
-      showCancelButton: false,
-      showCloseButton: false,
-      allowOutsideClick: false,
-      allowEscapeKey: false,
-      allowEnterKey: false,
-    });
-
-    const project = getCurrentProject();
-    project['name'] = projectName;
-
-    const data = new FormData();
-    data.append(utils.directoryTree.nodeTypes.PROJECT, JSON.stringify(project));
-    data.append('mode', mode);
-    data.append('path', path);
-
-    sendHttp('POST', 'saveProject', data, (request) => {
-      sweetAlert.close();
-
-      if (request.status !== 200) {
-        sweetAlert(
-          'Oops!',
-          'Could not save your project!!!  Please try again.',
-          'error'
-        );
-        return;
-      }
-
-      successFunc();
-      updateUI();
-    });
-  }
-
-  const selectedNode = utils.directoryTree.getSelectedNode();
-
-  if (
-    (selectedNode && projectName === selectedNode.name) ||
-    getNearestDirectory_().children.filter(
-      (node) => node.name === projectName && utils.directoryTree.isProject(node)
-    ).length === 0
-  ) {
-    go();
-  } else {
-    const msg = `${
-      'Are you sure you want to save over another project?\n\n' +
-      'The previous contents of '
-    }${projectName} will be permanently destroyed!`;
-    sweetAlert({
-      title: Alert.title('Warning'),
-      text: msg,
-      type: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#DD6B55',
-      confirmButtonText: 'Yes, overwrite it!',
-    }).then((result) => {
-      if (result && result.value) {
-        go();
-      }
-    });
-  }
-}
-
 function deleteProject_(path, buildMode, successFunc) {
-  const selectedNode = utils.directoryTree.getSelectedNode();
+  const selectedNode = DirTree.getSelectedNode();
   if (!selectedNode) return;
 
-  if (!signedIn()) {
+  if (!Auth.signedIn()) {
     sweetAlert('Oops', 'You must sign in to delete a project.', 'error');
     updateUI();
     return;
@@ -1229,8 +906,151 @@ function deleteProject_(path, buildMode, successFunc) {
   });
 }
 
+function saveProjectAsBase(successFunc, currentProject) {
+  if (!Auth.signedIn()) {
+    sweetAlert('Oops!', 'You must sign in to save files.', 'error');
+    updateUI();
+    return;
+  }
+
+  const selectedNode = DirTree.getSelectedNode();
+  const isDirectoryNode = DirTree.isDirectory(selectedNode);
+
+  sweetAlert({
+    title: Alert.title('Save As', 'mdi-cloud-upload'),
+    html:
+      selectedNode && isDirectoryNode
+        ? `Enter a name for your project in folder <b>${$('<div>')
+          .text(getNearestDirectory())
+          .html()
+          .replace(/ /g, '&nbsp;')}:`
+        : 'Enter a name for your project:',
+    input: 'text',
+    inputValue: selectedNode && !isDirectoryNode ? selectedNode.name : '',
+    confirmButtonText: 'Save',
+    showCancelButton: true,
+    closeOnConfirm: false,
+  }).then((result) => {
+    const parent = getNearestDirectory_();
+
+    function localSuccessFunc() {
+      const matches = parent.children.filter(
+        (node) => node.name === result.value && DirTree.isProject(node)
+      );
+      let node;
+      const type = DirTree.nodeTypes.PROJECT;
+      const name = result.value;
+
+      if (matches.length === 0) {
+        node = $('#directoryTree').tree(
+          'appendNode',
+          {
+            id: DirTree.createNodeId(type, name),
+            name,
+            type,
+            data: JSON.stringify(currentProject),
+          },
+          parent
+        );
+      } else {
+        node = matches[0];
+      }
+
+      $('#directoryTree').tree('selectNode', node);
+
+      updateChildrenIndexes(parent);
+
+      successFunc(result.value);
+    }
+    if (result && result.value) {
+      saveProjectBase(
+        getNearestDirectory(),
+        result.value,
+        window.projectEnv,
+        localSuccessFunc,
+        currentProject
+      );
+    }
+  });
+}
+
+function saveProjectBase(path, projectName, mode, successFunc, currentProject) {
+  if (!Auth.signedIn()) {
+    sweetAlert('Oops!', 'You must sign in to save files.', 'error');
+    updateUI();
+    return;
+  }
+
+  if (!projectName) return;
+
+  function go() {
+    sweetAlert({
+      title: Alert.title(`Saving ${projectName} ...`),
+      text: 'Saving your project. Please wait.',
+      showConfirmButton: false,
+      showCancelButton: false,
+      showCloseButton: false,
+      allowOutsideClick: false,
+      allowEscapeKey: false,
+      allowEnterKey: false,
+    });
+
+    const project = { ...currentProject };
+    project['name'] = projectName;
+
+    const data = new FormData();
+    data.append(DirTree.nodeTypes.PROJECT, JSON.stringify(project));
+    data.append('mode', mode);
+    data.append('path', path);
+
+    sendHttp('POST', 'saveProject', data, (request) => {
+      sweetAlert.close();
+
+      if (request.status !== 200) {
+        sweetAlert(
+          'Oops!',
+          'Could not save your project!!!  Please try again.',
+          'error'
+        );
+        return;
+      }
+
+      successFunc();
+      updateUI();
+    });
+  }
+
+  const selectedNode = DirTree.getSelectedNode();
+
+  if (
+    (selectedNode && projectName === selectedNode.name) ||
+    getNearestDirectory_().children.filter(
+      (node) => node.name === projectName && DirTree.isProject(node)
+    ).length === 0
+  ) {
+    go();
+  } else {
+    const msg = `${
+      'Are you sure you want to save over another project?\n\n' +
+      'The previous contents of '
+    }${projectName} will be permanently destroyed!`;
+    sweetAlert({
+      title: Alert.title('Warning'),
+      text: msg,
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#DD6B55',
+      confirmButtonText: 'Yes, overwrite it!',
+    }).then((result) => {
+      if (result && result.value) {
+        go();
+      }
+    });
+  }
+}
+
 function deleteFolder_(path, buildMode, successFunc) {
-  if (!signedIn()) {
+  if (!Auth.signedIn()) {
     sweetAlert('Oops', 'You must sign in to delete a folder.', 'error');
     updateUI();
     return;
@@ -1261,7 +1081,7 @@ function deleteFolder_(path, buildMode, successFunc) {
 
     sendHttp('POST', 'deleteFolder', data, (request) => {
       if (request.status === 200) {
-        const selectedNode = utils.directoryTree.getSelectedNode();
+        const selectedNode = DirTree.getSelectedNode();
         $('#directoryTree').tree('removeNode', selectedNode);
         successFunc();
         updateUI();
@@ -1272,7 +1092,7 @@ function deleteFolder_(path, buildMode, successFunc) {
 
 function createFolder(path, buildMode, successFunc) {
   warnIfUnsaved(() => {
-    if (!signedIn()) {
+    if (!Auth.signedIn()) {
       sweetAlert('Oops!', 'You must sign in to create a folder.', 'error');
       updateUI();
       return;
@@ -1312,21 +1132,21 @@ function createFolder(path, buildMode, successFunc) {
 
         successFunc();
 
-        let selectedNode = utils.directoryTree.getSelectedNode();
-        const type = utils.directoryTree.nodeTypes.DIRECTORY;
+        let selectedNode = DirTree.getSelectedNode();
+        const type = DirTree.nodeTypes.DIRECTORY;
         const name = result.value;
 
         if (!selectedNode) {
           selectedNode = $('#directoryTree').tree('getTree');
         }
-        if (selectedNode && !utils.directoryTree.isDirectory(selectedNode)) {
+        if (selectedNode && !DirTree.isDirectory(selectedNode)) {
           selectedNode = selectedNode.parent;
         }
 
         $('#directoryTree').tree(
           'appendNode',
           {
-            id: utils.directoryTree.createNodeId(type, name),
+            id: DirTree.createNodeId(type, name),
             name,
             type,
             children: [],
@@ -1341,7 +1161,7 @@ function createFolder(path, buildMode, successFunc) {
 }
 
 function loadProject_(path, name, buildMode, successFunc) {
-  if (!signedIn()) {
+  if (!Auth.signedIn()) {
     sweetAlert('Oops!', 'You must sign in to open projects.', 'error');
     updateUI();
     return;
@@ -1455,15 +1275,15 @@ function share() {
 }
 
 function shareFolder_(mode) {
-  if (!signedIn()) {
+  if (!Auth.signedIn()) {
     sweetAlert('Oops!', 'You must sign in to share your folder.', 'error');
     updateUI();
     return;
   }
 
-  const selectedNode = utils.directoryTree.getSelectedNode();
+  const selectedNode = DirTree.getSelectedNode();
 
-  if (!getNearestDirectory() || (selectedNode && selectedNode.name)) {
+  if (!getNearestDirectory() || !(selectedNode && selectedNode.name)) {
     sweetAlert('Oops!', 'You must select a folder to share!', 'error');
     updateUI();
     return;
@@ -1533,6 +1353,12 @@ function shareFolder_(mode) {
   });
 }
 
+function goto(line, col) {
+  codeworldEditor.getDoc().setCursor(line - 1, col - 1);
+  codeworldEditor.scrollIntoView(null, 100);
+  codeworldEditor.focus();
+}
+
 function preFormatMessage(msg) {
   while (msg.match(/(\r\n|[^\x08]|)\x08/)) {
     msg = msg.replace(/(\r\n|[^\x08])\x08/g, '');
@@ -1541,11 +1367,17 @@ function preFormatMessage(msg) {
   msg = Html.encode(msg)
     .replace(
       /program\.hs:(\d+):((\d+)(-\d+)?)/g,
-      '<a href="#" onclick="goto($1, $3); return false;">Line $1, Column $2</a>'
+      `<a href="#" onclick=${() => {
+        goto($1, $3);
+        return false;
+      }}>Line $1, Column $2</a>`
     )
     .replace(
       /program\.hs:\((\d+),(\d+)\)-\((\d+),(\d+)\)/g,
-      '<a href="#" onclick="goto($1, $2); return false;">Line $1-$3, Column $2-$4</a>'
+      `<a href="#" onclick=${() => {
+        goto($1, $2);
+        return false;
+      }}>Line $1-$3, Column $2-$4</a>`
     )
     .replace(
       /[A-Za-z0-9_-]*\.hs:(\d+):((\d+)(-\d+)?)/g,
@@ -1713,7 +1545,7 @@ function initDirectoryTree() {
     onCanMoveTo: (moving_node, target_node, position) => {
       // Forbid move inside project node,
       // but allow to move before and after
-      if (utils.directoryTree.isProject(target_node) && position === 'inside') {
+      if (DirTree.isProject(target_node) && position === 'inside') {
         return false;
       }
       if (target_node.type === 'loadNotification') return false;
@@ -1724,7 +1556,7 @@ function initDirectoryTree() {
     onCreateLi: function (node, $li) {
       const titleElem = $li.find('.jqtree-element .jqtree-title');
 
-      if (utils.directoryTree.isDirectory(node)) {
+      if (DirTree.isDirectory(node)) {
         if (node.is_open) {
           titleElem.before($('<i class="mdi mdi-18px mdi-folder-open"></i>'));
         } else {
@@ -1732,7 +1564,7 @@ function initDirectoryTree() {
         }
       } else if (node.type === 'loadNotification') {
         titleElem.before($('<div style="float: left" class="loader"></div>'));
-      } else if (utils.directoryTree.isProject(node)) {
+      } else if (DirTree.isProject(node)) {
         const asterisk = $('<i class="unsaved-changes"></i>');
         asterisk.css('display', 'none');
         titleElem.before($('<i class="mdi mdi-18px mdi-cube"></i>'));
@@ -1748,7 +1580,7 @@ function initDirectoryTree() {
   $('#directoryTree').on('tree.move', (event) => {
     event.preventDefault();
     warnIfUnsaved(() => {
-      if (!signedIn()) {
+      if (!Auth.signedIn()) {
         sweetAlert(
           'Oops!',
           'You must sign in to move this project or folder.',
@@ -1758,7 +1590,7 @@ function initDirectoryTree() {
         return;
       }
       const movedNode = event.move_info.moved_node;
-      const isFile = utils.directoryTree.isProject(movedNode);
+      const isFile = DirTree.isProject(movedNode);
       let fromPath, name;
       fromPath = pathToRootDir(movedNode);
       if (isFile) {
@@ -1800,7 +1632,7 @@ function initDirectoryTree() {
         if (haveChildWithSameNameAndType(movedNode, toNode)) {
           // Replacement of existing project
           let msg, confirmText;
-          if (utils.directoryTree.isProject(movedNode)) {
+          if (DirTree.isProject(movedNode)) {
             msg = `${
               'Are you sure you want to save over another project?\n\n' +
               'The previous contents of '
@@ -1837,7 +1669,7 @@ function initDirectoryTree() {
                   });
                   event.move_info.do_move();
                   updateChildrenIndexes(toNode);
-                  if (utils.directoryTree.isDirectory(movedNode)) {
+                  if (DirTree.isDirectory(movedNode)) {
                     loadSubTree(movedNode);
                     setCode('');
                   }
@@ -1886,7 +1718,7 @@ function initDirectoryTree() {
     event.preventDefault();
 
     const { node } = event;
-    const isProjectNode = utils.directoryTree.isProject(node);
+    const isProjectNode = DirTree.isProject(node);
 
     // Deselection of selected project. Cancel it and do nothing.
     if (isProjectNode && $('#directoryTree').tree('isNodeSelected', node)) {
@@ -1898,7 +1730,7 @@ function initDirectoryTree() {
 
         loadProject(node.name, path);
         $('#directoryTree').tree('selectNode', node);
-      } else if (utils.directoryTree.isDirectory(node)) {
+      } else if (DirTree.isDirectory(node)) {
         if (node.children.length === 0) {
           loadSubTree(node);
         }
@@ -1913,7 +1745,7 @@ function initDirectoryTree() {
 // Get directory nearest to selected node, or root if there is no selection
 function getNearestDirectory_(node) {
   if (node) {
-    const isDir = utils.directoryTree.isDirectory(node);
+    const isDir = DirTree.isDirectory(node);
     const hasParent = Boolean(node.parent);
 
     if (isDir) {
@@ -1925,14 +1757,14 @@ function getNearestDirectory_(node) {
     return node;
   }
 
-  const selectedNode = utils.directoryTree.getSelectedNode();
+  const selectedNode = DirTree.getSelectedNode();
 
   if (!selectedNode) {
     // nearest directory is root
     return $('#directoryTree').tree('getTree');
-  } else if (selectedNode && utils.directoryTree.isProject(selectedNode)) {
+  } else if (selectedNode && DirTree.isProject(selectedNode)) {
     return selectedNode.parent;
-  } else if (selectedNode && utils.directoryTree.isDirectory(selectedNode)) {
+  } else if (selectedNode && DirTree.isDirectory(selectedNode)) {
     return selectedNode;
   }
 }
@@ -1940,7 +1772,7 @@ function getNearestDirectory_(node) {
 function getNearestDirectory(node) {
   const selected = getNearestDirectory_(node);
   const path = pathToRootDir(selected);
-  if (utils.directoryTree.isDirectory(selected)) {
+  if (DirTree.isDirectory(selected)) {
     return path ? `${path}/${selected.name}` : selected.name;
   }
   return path;
@@ -1996,7 +1828,7 @@ function recalcChildrenIndexes(node) {
 }
 
 function updateChildrenIndexes(node) {
-  if (signedIn() && node && node.children) {
+  if (Auth.signedIn() && node && node.children) {
     recalcChildrenIndexes(node);
     const repacked = [];
     for (let i = 0; i < node.children.length; i++) {
@@ -2015,15 +1847,337 @@ function updateChildrenIndexes(node) {
 }
 
 function updateTreeOnNewProjectCreation() {
-  const selectedNode = getSelectedNode();
+  const selectedNode = DirTree.getSelectedNode();
 
-  if (selectedNode && utils.directoryTree.isProject(selectedNode)) {
-    utils.directoryTree.clearSelectedNode();
+  if (selectedNode && DirTree.isProject(selectedNode)) {
+    DirTree.clearSelectedNode();
 
     // Select parent folder (if any) as new project's location is
     // determined by the selection in the directory tree.
     if (pathToRootDir(selectedNode)) {
-      utils.directoryTree.selectNode(selectedNode.parent);
+      DirTree.selectNode(selectedNode.parent);
     }
   }
 }
+
+function clearWorkspace() {
+  DirTree.clearSelectedNode();
+
+  setCode('');
+}
+
+function isEditorClean() {
+  if (!window.codeworldEditor) {
+    return true;
+  }
+
+  const doc = window.codeworldEditor.getDoc();
+
+  if (window.savedGeneration === null) return doc.getValue() === '';
+  else return doc.isClean(window.savedGeneration);
+}
+
+/*
+ * Updates all UI components to reflect the current state.  The general pattern
+ * is to modify the state stored in variables and such, and then call updateUI
+ * to get the visual presentation to match.
+ */
+function updateUI() {
+  const selectedNode = DirTree.getSelectedNode();
+
+  const signOutButton = document.getElementById('signout');
+  const signInButton = document.getElementById('signin');
+  const navButton = document.getElementById('navButton');
+  const deleteButton = document.getElementById('deleteButton');
+  const saveButton = document.getElementById('saveButton');
+  const downloadButton = document.getElementById('downloadButton');
+  const shareFolderButton = document.getElementById('shareFolderButton');
+
+  if (Auth.signedIn()) {
+    if (signOutButton.style.display === 'none') {
+      signInButton.style.display = 'none';
+      signOutButton.style.display = '';
+      navButton.style.display = '';
+
+      window.mainLayout.show('west');
+      window.mainLayout.open('west');
+    }
+
+    if (selectedNode) {
+      deleteButton.style.display = '';
+    } else {
+      deleteButton.style.display = 'none';
+    }
+
+    if (selectedNode && DirTree.isProject(selectedNode)) {
+      saveButton.style.display = '';
+      shareFolderButton.style.display = 'none';
+
+      if (downloadButton) {
+        downloadButton.style.display = '';
+      }
+    } else if (selectedNode && DirTree.isDirectory(selectedNode)) {
+      saveButton.style.display = 'none';
+      shareFolderButton.style.display = '';
+
+      if (downloadButton) {
+        downloadButton.style.display = 'none';
+      }
+    } else {
+      saveButton.style.display = 'none';
+      shareFolderButton.style.display = 'none';
+
+      if (downloadButton) {
+        downloadButton.style.display = 'none';
+      }
+    }
+  } else {
+    if (signOutButton.style.display === '') {
+      signInButton.style.display = '';
+      signOutButton.style.display = 'none';
+      saveButton.style.display = 'none';
+
+      window.mainLayout.hide('west');
+    }
+
+    navButton.style.display = 'none';
+    deleteButton.style.display = 'none';
+    shareFolderButton.style.display = 'none';
+  }
+
+  const inspectButton = document.getElementById('inspectButton');
+
+  if (inspectButton) {
+    if (window.debugAvailable) {
+      inspectButton.style.display = '';
+
+      if (window.debugActive) {
+        inspectButton.style.color = 'black';
+      } else {
+        inspectButton.style.color = '';
+      }
+    } else {
+      inspectButton.style.display = 'none';
+    }
+  }
+
+  document.getElementById('newButton').style.display = '';
+  document.getElementById('saveAsButton').style.display = '';
+  document.getElementById('runButtons').style.display = '';
+
+  let title = selectedNode ? selectedNode.name : '(new)';
+
+  if (!isEditorClean()) {
+    title = `* ${title}`;
+
+    if (selectedNode && DirTree.isProject(selectedNode)) {
+      const asterisk = selectedNode.element.getElementsByClassName(
+        'unsaved-changes'
+      )[0];
+      if (asterisk) {
+        asterisk.style.display = '';
+      }
+    }
+  } else {
+    $('.unsaved-changes').css('display', 'none');
+  }
+
+  // If true - code currently in document is not equal to
+  // last compiled code
+  const running = document.getElementById('runner').style.display !== 'none';
+  const obsolete = window.codeworldEditor
+    ? !window.codeworldEditor.getDoc().isClean(window.runningGeneration)
+    : false;
+  const obsoleteAlert = document.getElementById('obsolete-code-alert');
+  if (running && obsolete) {
+    obsoleteAlert.classList.add('obsolete-code-alert-fadein');
+    obsoleteAlert.classList.remove('obsolete-code-alert-fadeout');
+  } else {
+    obsoleteAlert.classList.add('obsolete-code-alert-fadeout');
+    obsoleteAlert.classList.remove('obsolete-code-alert-fadein');
+  }
+
+  document.title = `${title} - CodeWorld`;
+}
+
+function run(hash, dhash, msg, error, generation) {
+  window.runningGeneration = generation;
+  window.debugAvailable = false;
+  window.debugActive = false;
+  window.lastRunMessage = msg;
+
+  const runner = document.getElementById('runner');
+
+  // Stop canvas recording if the recorder is active
+  document.getElementById('runner').contentWindow.postMessage(
+    {
+      type: 'stopRecord',
+    },
+    '*'
+  );
+
+  if (hash) {
+    window.location.hash = `#${hash}`;
+    document.getElementById('shareButton').style.display = '';
+  } else {
+    window.location.hash = '';
+    document.getElementById('shareButton').style.display = 'none';
+  }
+
+  if (dhash) {
+    const loc = `run.html?dhash=${dhash}&mode=${window.buildMode}`;
+    runner.contentWindow.location.replace(loc);
+    if (
+      Boolean(navigator.mediaDevices) &&
+      Boolean(navigator.mediaDevices.getUserMedia)
+    ) {
+      document.getElementById('startRecButton').style.display = '';
+    }
+  } else {
+    runner.contentWindow.location.replace('about:blank');
+    document.getElementById('runner').style.display = 'none';
+    document.getElementById('startRecButton').style.display = 'none';
+  }
+
+  if (hash || msg) {
+    window.mainLayout.show('east');
+    window.mainLayout.open('east');
+    document.getElementById('shareFolderButton').style.display = 'none';
+  } else {
+    document.getElementById('shareFolderButton').style.display = '';
+    window.mainLayout.hide('east');
+  }
+
+  clearMessages();
+
+  parseCompileErrors(msg).forEach((cmError) => {
+    printMessage(cmError.severity, cmError.fullText);
+  });
+
+  if (error) markFailed();
+
+  window.deployHash = dhash;
+
+  updateUI();
+  document.getElementById('runner').addEventListener('load', updateUI);
+}
+
+function stopRun() {
+  if (window.debugActive) {
+    document.getElementById('runner').contentWindow.postMessage(
+      {
+        type: 'stopDebug',
+      },
+      '*'
+    );
+    TreeDialog.destroy();
+  }
+  window.cancelCompile();
+
+  run('', '', '', false, null);
+}
+
+function parseCompileErrors(rawErrors) {
+  const errors = [];
+  rawErrors = rawErrors.split('\n\n');
+  rawErrors.forEach((err) => {
+    const lines = err.trim().split('\n');
+    const firstLine = lines[0].trim();
+    const otherLines = lines
+      .slice(1)
+      .map((ln) => ln.trim())
+      .join('\n');
+    const re1 = /^program\.hs:(\d+):((\d+)-?(\d+)?): (\w+):(.*)/;
+    const re2 = /^program\.hs:\((\d+),(\d+)\)-\((\d+),(\d+)\): (\w+):(.*)/;
+
+    if (err.trim() === '') {
+      // Ignore empty messages.
+    } else if (re1.test(firstLine)) {
+      const match = re1.exec(firstLine);
+
+      const line = Number(match[1]) - 1;
+      let startCol = Number(match[3]) - 1;
+      let endCol;
+      if (match[4]) {
+        endCol = Number(match[4]);
+      } else {
+        const token = window.codeworldEditor
+          .getLineTokens(line)
+          .find((t) => t.start === startCol);
+        if (token) {
+          endCol = token.end;
+        } else if (
+          startCol >= window.codeworldEditor.getDoc().getLine(line).length
+        ) {
+          endCol = startCol;
+          --startCol;
+        } else {
+          endCol = startCol + 1;
+        }
+      }
+
+      const message = ((match[6] ? `${match[6].trim()}\n` : '') + otherLines)
+        .replace(/program\.hs:(\d+):((\d+)(-\d+)?)/g, 'Line $1, Column $2')
+        .replace(
+          /program\.hs:\((\d+),(\d+)\)-\((\d+),(\d+)\)/g,
+          'Line $1-$3, Column $2-$4'
+        );
+
+      errors.push({
+        from: CodeMirror.Pos(line, startCol),
+        to: CodeMirror.Pos(line, endCol),
+        severity: match[5],
+        fullText: err,
+        message: message,
+      });
+    } else if (re2.test(firstLine)) {
+      const match = re2.exec(firstLine);
+
+      const startLine = Number(match[1]) - 1;
+      const startCol = Number(match[2]) - 1;
+      const endLine = Number(match[3]) - 1;
+      const endCol = Number(match[4]);
+
+      errors.push({
+        from: CodeMirror.Pos(startLine, startCol),
+        to: CodeMirror.Pos(endLine, endCol),
+        severity: match[5],
+        fullText: err,
+        message: (match[6] ? `${match[6].trim()}\n` : '') + otherLines,
+      });
+    } else {
+      console.log('Can not parse error header:', firstLine);
+    }
+  });
+  return errors;
+}
+
+export {
+  clearMessages,
+  clearWorkspace,
+  createFolder,
+  definePanelExtension,
+  deleteFolder_,
+  deleteProject_,
+  discoverProjects,
+  getNearestDirectory,
+  initDirectoryTree,
+  isEditorClean,
+  loadProject_,
+  markFailed,
+  onHover,
+  parseCompileErrors,
+  parseSymbolsFromCurrentCode,
+  printMessage,
+  registerStandardHints,
+  run,
+  saveProjectBase,
+  saveProjectAsBase,
+  setCode,
+  share,
+  shareFolder_,
+  stopRun,
+  updateTreeOnNewProjectCreation,
+  updateUI,
+  warnIfUnsaved,
+};
