+++ web/js/utils/html.js
@@ -1,0 +1,5 @@
+function encode(str) {
+  return $('<div>').text(str).html();
+}
+
+export { encode };
