--- web/js/utils/index.js
@@ -1,3 +1,0 @@
-'use strict';
-
-window.utils = {};
